Refactor packages: controller package removed, ZooManager class moved to new animals package. model package renamed to domain and added into the new animals package.
